docs(button): update documentation for component structure after circular dependency fix
