+++ packages/ui/src/components/Button/Button.types.ts
@@ -1,6 +1,6 @@
 import { type VariantProps } from "class-variance-authority";
 import React from "react";
-import { buttonVariants } from "./Button.view";
+import { buttonVariants } from "./Button.variants";
 
 export interface ButtonProps
 	extends React.ButtonHTMLAttributes<HTMLButtonElement>,
