+++ packages/ui/src/components/Button/Button.view.tsx
@@ -1,37 +1,8 @@
-import { cva } from "class-variance-authority";
 import { Loader2 } from "lucide-react";
 import React from "react";
 import { cn } from "../../lib/utils";
 import { ButtonProps } from "./Button.types";
-
-export const buttonVariants = cva(
-	"inline-flex items-center justify-center rounded-md text-sm font-medium transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:opacity-50 disabled:pointer-events-none ring-offset-background",
-	{
-		variants: {
-			variant: {
-				default: "bg-primary text-primary-foreground hover:bg-primary/90",
-				destructive:
-					"bg-destructive text-destructive-foreground hover:bg-destructive/80",
-				outline:
-					"border border-border bg-transparent hover:bg-accent hover:text-accent-foreground",
-				secondary:
-					"bg-secondary text-secondary-foreground hover:bg-secondary/80",
-				ghost: "hover:bg-accent hover:text-accent-foreground",
-				link: "text-foreground underline-offset-4 hover:underline hover:text-primary",
-			},
-			size: {
-				default: "h-10 py-2 px-4",
-				sm: "h-9 px-3 rounded-md",
-				lg: "h-11 px-8 rounded-md",
-			},
-		},
-		defaultVariants: {
-			variant: "default",
-			size: "default",
-		},
-	},
-);
-
+import { buttonVariants } from "./Button.variants";
 export const ButtonView = React.forwardRef<HTMLButtonElement, ButtonProps>(
 	({ className, children, variant, size, isLoading, ...props }, ref) => {
 		return (
